Moved FgsMb2 to FgsMb, adjusted code instances to match. Fixed tests.
